Parameterize subsystem test case.
